chore(analytics): Create a new test target for Analytics stress tests
